Move types around.
